Split createSuperContext function to its own file. Minor cleanup across project.
