+++ package.json
@@ -28,7 +28,7 @@
     "@types/react-dom": "^16.9.8",
     "@typescript-eslint/eslint-plugin": "4.5.0",
     "@typescript-eslint/parser": "4.5.0",
-    "eslint": "7.11.0",
+    "eslint": "6.8.0",
     "eslint-config-prettier": "6.13.0",
     "eslint-config-react-app": "5.2.1",
     "eslint-plugin-prettier": "^3.1.4",
