+++ src/SuperContext.tsx
@@ -1,23 +1,21 @@
-import React, { Context, createContext, PropsWithChildren, useContext, useEffect } from "react";
-import { SuperContextInterceptor } from "./Interceptors";
-
-export interface SuperContextDefinition<P = any, T = any> {
-  context: Context<T>;
-  factory: (props: P) => T;
-  props: P;
-  options?: Partial<CreateSuperContextOptions<T>>;
-}
-
-export interface CreateSuperContextOptions<T> {
-  displayName: string;
-  interceptors: SuperContextInterceptor<T>[];
-}
+import React, { PropsWithChildren, useEffect, useMemo } from "react";
+import { CreateSuperContextOptions, SuperContextDefinition } from "./CreateSuperContext";
 
 export type SuperContextProps = PropsWithChildren<{
+  /**
+   * Contexts provided by the SuperContext.
+   */
   contexts: (SuperContextDefinition | ((props?: any) => SuperContextDefinition))[];
+  /**
+   * Options to apply to all contexts provided by the SuperContext. Will be overwritten by any
+   * context specific options.
+   */
   defaultOptions?: Partial<CreateSuperContextOptions<any>>;
 }>;
 
+/**
+ * The SuperContext is a provider component for all contexts created with createSuperContext.
+ */
 export const SuperContext = ({
   contexts,
   defaultOptions,
@@ -26,12 +24,18 @@ export const SuperContext = ({
   contexts.length === 0 ? (
     <>{children}</>
   ) : (
-    <SuperContextProvider contexts={contexts} defaultOptions={defaultOptions} index={0}>
+    <SubContext contexts={contexts} defaultOptions={defaultOptions} index={0}>
       {children}
-    </SuperContextProvider>
+    </SubContext>
   );
 
-const SuperContextProvider = ({
+// library's default options, can be overridden
+const fallbackOptions: CreateSuperContextOptions<any> = {
+  interceptors: [],
+  displayName: "SuperContext",
+};
+
+const SubContext = ({
   contexts,
   defaultOptions,
   index,
@@ -43,50 +47,38 @@ const SuperContextProvider = ({
   const { context, factory, props, options }: SuperContextDefinition =
     typeof superContext === "function" ? superContext() : superContext;
 
+  const { displayName, interceptors } = useMemo<CreateSuperContextOptions<any>>(
+    () => ({
+      ...fallbackOptions, // library defaults
+      ...defaultOptions, // defaults of the super context
+      ...options, // options of the sub-context
+    }),
+    [defaultOptions, options]
+  );
+
   let value = factory(props);
 
-  const interceptors = options?.interceptors ?? defaultOptions?.interceptors;
   if (interceptors) {
-    for (const plugin of interceptors) {
-      value = plugin(value);
+    for (const interceptor of interceptors) {
+      value = interceptor(value);
     }
   }
 
   useEffect(() => {
-    context.displayName = options?.displayName ?? defaultOptions?.displayName ?? "SuperContext";
-  }, [context, options?.displayName, defaultOptions?.displayName]);
+    context.displayName = displayName;
+  }, [displayName]);
 
   const nextIndex = index + 1;
 
   return (
     <context.Provider value={value}>
       {nextIndex < contexts.length ? (
-        <SuperContextProvider contexts={contexts} index={nextIndex}>
+        <SubContext contexts={contexts} index={nextIndex}>
           {children}
-        </SuperContextProvider>
+        </SubContext>
       ) : (
         children
       )}
     </context.Provider>
   );
 };
-
-// could maybe benefit from partial type argument inference: https://github.com/microsoft/TypeScript/issues/26242
-export function createSuperContext<T, P = any>(
-  factory: (props: P) => T,
-  options: Partial<CreateSuperContextOptions<T>> = {}
-): [(props: P) => SuperContextDefinition<P, T>, () => T] {
-  const defaultValue = {} as T;
-  const context = createContext<T>(defaultValue);
-  const useSuperContext = () => {
-    const value = useContext<T>(context);
-    if (value === defaultValue) {
-      const forName = options?.displayName ? `(for ${options.displayName}) ` : "";
-      throw new Error(
-        `Super context hook ${forName}was called, but the context was not provided. Make sure to include the context to a SuperContext.`
-      );
-    }
-    return value;
-  };
-  return [(props: P) => ({ context, factory, props, options }), useSuperContext];
-}
