+++ src/index.ts
@@ -1,2 +1,3 @@
+export * from "./CreateSuperContext";
 export * from "./SuperContext";
 export * from "./Interceptors";
